router/requirements.txt
aiohttp
google-genai
openai
groq
python-dotenv
pymilvus
